Update with Getx
