--- task_application/lib/Project Page/components/models/laptops.dart
@@ -1,8 +1,0 @@
-class Laptops {
-  final String name;
-  final String price;
-  final String image;
-  final String description;
-
-  Laptops({required this.name, required this.price, required this.image, required this.description});
-}
