--- task_application/lib/Project Page/pages/ordersuccesspage.dart
@@ -1,32 +1,0 @@
-import 'package:flutter/material.dart';
-
-class OrderSuccessPage extends StatelessWidget {
-  const OrderSuccessPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Order Success'),
-      ),
-      body: const Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              Icons.check_circle_outline,
-              color: Colors.green,
-              size: 100,
-            ),
-            SizedBox(height: 20),
-            Text(
-              'Your order has been placed successfully!',
-              style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
