+++ task_application/lib/models/task.dart
@@ -1,0 +1,15 @@
+class Task {
+  String title;
+  String description;
+  DateTime creationDate;
+  bool isCompleted;
+
+  Task({
+    required this.title,
+    required this.description,
+    required this.creationDate,
+    this.isCompleted = false,
+  });
+
+  bool get completed => isCompleted;
+}
