+++ task_application/pubspec.yaml
@@ -1,4 +1,4 @@
-name: task_application
+name: taskapp_v2
 description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.4.1 <4.0.0'
+  sdk: '>=3.4.3 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,26 +30,28 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_slidable: 3.1.0
-  dots_indicator: ^3.0.0
-  flutter_speed_dial: ^7.0.0
+    
+    
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
+  flutter_speed_dial: ^7.0.0
+  get: ^4.6.6
+  dots_indicator: ^3.0.0
+  flutter_slidable: ^3.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^4.0.0
+  flutter_lints: ^3.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -64,16 +66,8 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-    - assets/images/logo.jpg
-    - assets/images/dev-shops_logo.png
-    - assets/images/apple.webp
-    - assets/images/alienware.jpg
-    - assets/images/dell.webp
-    - assets/images/laptop.png
-    - assets/images/samsung.webp
-    - assets/images/omen.webp
-    - assets/images/microsoft.png
     - assets/undraw.png
+  #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
@@ -87,17 +81,16 @@ flutter:
   # list giving the asset and other descriptors for the font. For
   # example:
   # fonts:
-    # - family: Schyler
-    #   fonts:
-    #     - asset: fonts/Schyler-Regular.ttf
-    #     - asset: fonts/Schyler-Italic.ttf
-    #       style: italic
-    # - family: Product Sans
-      # fonts: 
-        # - asset: fonts/ProductSans-Black.ttf
-        # - asset: fonts/Product-Bold.ttf
-        # - asset: fonts/ProductSans-Regular.ttf
-          # weight: 400
-  
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
